+++ CHANGELOG.md
@@ -26,7 +26,23 @@ v1.0.0-alpha2
   ApiOpenStudio (see issue #54 in ApiOpenStudio GitLab repo).
 - License updates.
 
-v1.0.0-beta1 (in progress)
-==========================
+v1.0.0-alpha3
+=============
 
-- Updated notes for contributing
+- Gulpfile fixes.
+- README.md update.
+- Updated authentication to use the new JWT aut.
+- Updated handling of responses for:
+  - vars
+  - users
+  - roles
+  - resources
+
+v1.0.0-beta
+===========
+
+- Added OpenApi UI and editor.
+- Updated all API calls the handle possible new JSON response objects
+  (see https://gitlab.com/apiopenstudio/apiopenstudio_admin/-/issues/23):
+  - New JSON error response object.
+  - Responses can now be raw JSON response or a JSON response object.